Properly handle negative balanaces in the Electrum RPC
Refs https://github.com/bitcoindevkit/rust-electrum-client/issues/45

+++ src/electrum/server.rs
@@ -286,8 +286,8 @@ impl Connection {
         let (chain_stats, mempool_stats) = self.query.stats(&script_hash[..]);
 
         Ok(json!({
-            "confirmed": chain_stats.funded_txo_sum - chain_stats.spent_txo_sum,
-            "unconfirmed": mempool_stats.funded_txo_sum - mempool_stats.spent_txo_sum,
+            "confirmed": chain_stats.funded_txo_sum as i64 - chain_stats.spent_txo_sum as i64,
+            "unconfirmed": mempool_stats.funded_txo_sum as i64 - mempool_stats.spent_txo_sum as i64,
         }))
     }
